+++ src/main/java/com/code/pojo/Teacher.java
@@ -1,12 +1,12 @@
 package com.code.pojo;
 
 import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.Version;
 import com.baomidou.mybatisplus.annotation.TableId;
-import java.io.Serializable;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
+
+import java.io.Serializable;
 
 /**
  * <p>
@@ -24,7 +24,7 @@ public class Teacher implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "teacher_id", type = IdType.AUTO)
-    private Integer teacherId;
+    private Long teacherId;
 
     private String teacherNo;
